refactor(tab): split tabbar.tsx into focused tab-*.ts modules
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Bundled core tab-strip render/select/close, reorder drag-and-drop, tear-off
RPC orchestration, a large cross-window tear-off event-listener subsystem
(Phase 4/5 IPC handlers), and active-tab-color-line measurement. Split into
tab-close-confirm-modal.tsx, tab-tearoff-rpc.ts, tab-tearoff-events.ts
(useTabTearOffEvents), tab-reorder.ts (useTabDragAndDrop), tab-color-line.ts
(useActiveTabColorLine). Shared drag state continues to live in the
untouched tabbar-dnd.ts, read (not duplicated) by the new files. The three
shared DOM refs (tabBarRef/tabBarScrollRef/tabBarFillRef) are threaded into
each hook as zero-arg accessor closures, matching the codebase's existing
pattern for lazily-dereferenced JSX refs.

1366 -> 263 lines. tsc --noEmit clean; tabbar-dnd.test.ts (38 tests) passes
unchanged.

<!-- agentmux:agent_id=agentx -->
